services/user/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
motor==3.3.2
pydantic==2.5.0
PyJWT==2.8.0
bcrypt==4.1.1
